@@ -1,11 +1,10 @@
 /*!
- * Copyright (c) 2021 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2021-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
+import * as bedrock from '@bedrock/core';
+import '@bedrock/views';
+import '@bedrock/webpack';
 
-const bedrock = require('bedrock');
-require('bedrock-views');
-require('bedrock-webpack');
-require('./config');
+import './config.js';
 
 bedrock.start();
